refactor: inline chat handler & code action handler
