+++ packages/startup/entry/sample-modules/ai-native/ai-native.contribution.ts
@@ -7,8 +7,8 @@ import {
   NodeMatcher,
   ShellMatcher,
   TSCMatcher,
-} from '@opensumi/ide-ai-native/lib/browser/terminal/matcher';
-import { TextWithStyle } from '@opensumi/ide-ai-native/lib/browser/terminal/utils/ansi-parser';
+} from '@opensumi/ide-ai-native/lib/browser/contrib/terminal/matcher';
+import { TextWithStyle } from '@opensumi/ide-ai-native/lib/browser/contrib/terminal/utils/ansi-parser';
 import {
   AINativeCoreContribution,
   IChatFeatureRegistry,
@@ -26,19 +26,15 @@ import { TerminalDetectionPromptManager } from '@opensumi/ide-ai-native/lib/comm
 import { Domain, getIcon } from '@opensumi/ide-core-browser';
 import {
   AIBackSerivcePath,
-  AbortError,
   CancelResponse,
   ChatServiceToken,
   ErrorResponse,
   IAIBackService,
-  IChatContent,
-  IChatProgress,
   MergeConflictEditorMode,
   ReplyResponse,
   getDebugLogger,
 } from '@opensumi/ide-core-common';
 import { ICodeEditor, NewSymbolName, NewSymbolNameTag } from '@opensumi/ide-monaco';
-import { listenReadable } from '@opensumi/ide-utils/lib/stream';
 import { MarkdownString } from '@opensumi/monaco-editor-core/esm/vs/base/common/htmlContent';
 
 import { SlashCommand } from './SlashCommand';
